+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## UNRELEASE
 
 - **Fix** : (OneClick) Prevent saving payment card details without proper authorization.
+- **Fix** : Remove SDK js init from email template Multibanco/Sisal
 
 ## 2.7.1